+++ CHANGES.md
@@ -5,6 +5,11 @@
 - Added Django 5.0 support
 - Multiple maintenance improvements
 
+**2.1.1**
+
+- 100% coverage and bugfix in edge case
+- Updates via ambient package updater
+
 **2.1.0**
 
 - Added Python 3.12 support